Refactoring for a cleaner api interface
