+++ pyorthanc/__init__.py
@@ -1,16 +1,17 @@
+from . import errors, util
+from ._filtering import build_patient_forest, find, trim_patients
+from ._find import find_instances, find_patients, find_series, find_studies, query_orthanc
+from ._modality import Modality, RemoteModality
+from ._resources import Instance, Patient, Series, Study
 from .async_client import AsyncOrthanc
 from .client import Orthanc
-from .filtering import build_patient_forest, find, trim_patients
-from .find import find_instances, find_patients, find_series, find_studies, query_orthanc
-from .remote import RemoteModality
-from .resources import Instance, Patient, Series, Study
 from .retrieve import retrieve_and_write_instance, retrieve_and_write_patient, retrieve_and_write_patients, \
     retrieve_and_write_series, retrieve_and_write_study
-
 
 __all__ = [
     'AsyncOrthanc',
     'Orthanc',
+    'Modality',
     'RemoteModality',
     'Patient',
     'Study',
@@ -29,4 +30,6 @@ __all__ = [
     'retrieve_and_write_study',
     'retrieve_and_write_series',
     'retrieve_and_write_instance',
+    'util',
+    'errors',
 ]
