+++ pyorthanc/_resources/resource.py
