+++ pyorthanc/_resources/series.py
